requirements.txt
cleval
google-cloud-documentai
google-cloud-core
google-cloud-vision